+++ CHANGELOG.md
@@ -33,6 +33,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Maintenance
    * Added missing unit tests for `pysat.utils.time`
    * Added missing unit tests for `pysat.utils.file.parse_delimited_filename`
+   * Streamlined unit tests for `test_orbits`
 
 [3.0.1] - 2021-XX-XX
 --------------------